Use migration pagination package
React Bootstrap Pagination API changed a lot in 0.32, and they released
a package with the old API that can be used while we migrate the code.
Since that API is likely to change in React Bootstrap 1 and Bootstrap 4,
I think it is fine for now to use the old API and make any changes
later.